requirements.txt
asgiref==3.8.1
boto3==1.34.126
botocore==1.34.126
crispy-bootstrap5==2024.2
Django==5.0.6
django-crispy-forms==2.1
django-environ==0.11.2
django-storages==1.14.3
gunicorn==22.0.0
jmespath==1.0.1
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
